respond to cmcewen comments
